project/requirement.txt
asgiref==3.4.1
Django==3.2.10
django-cors-headers==3.10.1
djangorestframework==3.13.0
Pillow==8.4.0
pytz==2021.3
sqlparse==0.4.2
typing_extensions==4.0.1
